Add more OOP
Created 4 extra classes,
ui, tasklist, storage and parser
