+++ data/duke.txt
@@ -6,7 +6,4 @@ T|1|pray
 T|0|pray
 D|0|deadline dota |by 2 of February 2024, 12:00AM
 E|0|event project meeting |2-2-2024|4-2-2024
-E|0|event halflife |2/2/2024 1800|5/2/2024 1900
-D|0|deadline games |by 2 of February 2024, 6:00PM
-E|0|event halflife |2/2/2024 1800|5/2/2024 1900
-E|0|event halflife |2 of February 2024, 6:00PM|5 of February 2024, 7:00PM
+T|1|study
