+++ src/main/java/Ui.java
@@ -1,0 +1,11 @@
+public class Ui {
+    public void showLoadingError() {
+        System.out.println("Error loading tasks from file.");
+    }
+
+    public void showByeMessage() {
+        System.out.println("Bye. Hope to see you again soon!");
+    }
+
+    // Add other UI-related methods as needed
+}
